pyproject.toml
[project]
name = "repl-nix-finsmartai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.2.0",
    "numpy>=2.1.2",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "streamlit>=1.39.0",
    "stripe>=11.1.1",
]
